requirements.txt
# TTS
fastapi[standard]
coqui-tts[server,pt]
pydantic
soundfile
# python-keycloak
# fastapi-keycloak-middleware
# fastapi-users[sqlalchemy,oauth]
# uvicorn[standard]
python-keycloak
jinja2